Rename io.bitsquare.{di=>}AbstractBitsquareModule
