+++ src/main/java/io/bitsquare/btc/BitcoinModule.java
@@ -17,7 +17,7 @@
 
 package io.bitsquare.btc;
 
-import io.bitsquare.di.AbstractBitsquareModule;
+import io.bitsquare.AbstractBitsquareModule;
 
 import org.bitcoinj.core.NetworkParameters;
 import org.bitcoinj.params.MainNetParams;
