+++ src/main/java/io/bitsquare/crypto/CryptoModule.java
@@ -17,7 +17,7 @@
 
 package io.bitsquare.crypto;
 
-import io.bitsquare.di.AbstractBitsquareModule;
+import io.bitsquare.AbstractBitsquareModule;
 
 import java.util.Properties;
